feat(courrier-service): orgainize function to separate utils
and the test case
